part1 final commit

+++ project1/part1/logger.cpp
@@ -34,7 +34,7 @@ void Logger::printLog()
   printf("***************** Printing Vector Unit Execution Log *****************\n");
   printf(" Instruction | Vector Lane Occupancy ('*' for active, '_' for inactive)\n");
   printf("------------- --------------------------------------------------------\n");
-  for (int i = 0; i < log.size(); i++)
+  for (size_t i = 0; i < log.size(); i++)
   {
     printf("%12s | ", log[i].instruction);
     for (int j = 0; j < VECTOR_WIDTH; j++)
